feat(event): implement InitEvent and MoveEvent as subtypes
